requirements.txt
tabulate
matplotlib
python-dotenv
mysql-connector-python
python-dotenv
pyfiglet